+++ tests/processtests.cpp
@@ -16,7 +16,7 @@ std::unique_ptr<Process> ProcessTest::m_proc{ nullptr };
 TEST_F(ProcessTest, Create)
 {
 #ifdef _WIN32
-    ASSERT_NO_THROW(m_proc = std::make_unique<Process>(Environment::findDependency("cmd.exe"), std::vector<std::string>{ "/c", "timeout", "60" }));
+    ASSERT_NO_THROW(m_proc = std::make_unique<Process>(Environment::findDependency("cmd.exe"), std::vector<std::string>{ "/c", "start", "/min", "timeout.exe", "60" }));
 #else
     ASSERT_NO_THROW(m_proc = std::make_unique<Process>(Environment::findDependency("sleep"), std::vector<std::string>{ "60" }));
 #endif
@@ -43,4 +43,18 @@ TEST_F(ProcessTest, Kill)
 TEST_F(ProcessTest, Destroy)
 {
     ASSERT_NO_THROW(m_proc.reset());
+}
+
+TEST_F(ProcessTest, Send)
+{
+#ifdef _WIN32
+    Process p{ Environment::findDependency("cmd.exe") };
+#else
+    Process p{ Environment::findDependency("sh") };
+#endif
+    ASSERT_TRUE(p.start());
+    ASSERT_TRUE(p.sendCommand("echo \"Hello\""));
+    ASSERT_TRUE(p.sendCommand("exit"));
+    p.waitForExit();
+    ASSERT_FALSE(p.getOutput().empty());
 }